Update Dockerfile for updated Browsh and Firefox
The old Dockerfile uses Debian Stretch as a base image. Firefox from Stretch is quite old and some modern webapps are broken on it.
Fixing the Dockerfile to use Debian Bullseye as a base image, plus the latest (v1.8.0) browsh image, lets the container use the latest stable Firefox version (107 at the time of writing this, which works perfectly).